NSUrlSessionHandler - Support basic auth and multiple auth headers (#5052)
If the challenge is for basic authentication then try to get the
credentials for this authentication type.

Also handle multiple WWW-Authenticate headers, such as:

WWW-Authenticate: Negotiate,  NTLM

If the credentials provided on the NSUrlSessionHandler are not
available for a specific authentication type, or if the authentication
method is not supported, then reject the protection space. This allows
the next authentication challenge to be sent to the DidReceiveChallenge
method. Also handle a null reference exception that could occur if
there were no credentials available for the authentication type.