[msbuild] Add CreateBindingResourcePackage task
- The Xamarin.Shared logic depends upon 2 msbuild tasks, the first one being CreateBindingResourcePackage
- This task is executed after compilation of binding projects, and has three jobs
   - Create a "resource package" folder next to the output assembly foo.resources
   - Copy the native reference(s) into a that folder to be found later
   - Create a manifest xml file with two major sections
      - The MVID of the binding assembly just built which will be double chcked later at consumption
      - A serialized representation of each msbuild native reference which will be rehydrated later

Do note, that there are two use cases we must handle:
 - In solution: The binding project and the consuming application are in the same build process
 - Out of solution: The binding project is built and the output is provided as .dll files referenced in a seperate csproj later

This binding resource bundle is needed to handle the "out of solution" use case.

Locally this adds about 150ms to binding project builds, so it can be enabled by default in projects
that opt-in to the new NoBindingEmbedding style binding projects. An escape hatch of <NoBindingEmbedding> has been added just in case.