Add config file for libmono-native dllmap
The sn tool is used during the build to sign assemblies, but with a twist:
It uses the MONO_CFG_DIR environment variable to force a custom machine.config
from the 4.5/mono/machine.config directory with custom key mappings.

That in turn also changes where the global dllmap file is looked up though
and now that certain core parts of the BCL rely on System.Native we need
to make sure we have it in place.