[Foundation] Ensure that Connection Losts exceptions are more reliable.
In some cases, when the connection is lost, the Stream is not set to
have an exception and continues the read as normal. We need to ensure
that the stream cannot be read when setting the response to an
exception. We block the reading, set both exceptions and continue. In
that case, the exception will be correctly raised when trying to read
from the stream.

A new check has to be done after the clean up of the stream if the
session is using caching.

Fixes https://github.com/xamarin/xamarin-macios/issues/5215