[Foundation] Ensure that Connection Losts exceptions are more reliable. (#5769)
In some cases, when the connection is lost, the Stream is not set to
have an exception and continues the read as normal. We need to ensure
that the stream cannot be read when setting the response to an
exception. 

In order to do that, we use a new CancellationTokenSource whose token will be used for the HttpContent operations, therefore, if we received an error in the delegate, we can cancel the appropriate stream operations.

A new check has to be done after the clean up of the stream if the
session is using caching.

Fixes https://github.com/xamarin/xamarin-macios/issues/5215