[generator] Preserve the `Handler` fields inside generated trampolines code
This is caused by a new optimization by the linker which makes it
possible to remove some `.cctor` if it's fields are not preserved.

In earlier versions the linker would actually mark the fields because
of the `.cctor` (whenever a type is used).

While this is the correct fix we'll also need (in progress) to allow
disabling this optimization for existing _linker safe_ code (in part
because other, similar pattern might exists in 3rd party code).

Fix failures such as
```
[FAIL] RegistrarTest.BlockReturnTest : ObjCRuntime.RuntimeException : Invalid DelegateProxyAttribute for the return value for the method MonoTouchFixtures.ObjCRuntime.RegistrarTest+BlockReturnTestClass.MethodReturningBlock: DelegateType (ObjCRuntime.Trampolines+SDRegistrarTestBlock) specifies a type without a 'Handler' field. Please file a bug at https://github.com/xamarin/xamarin-macios/issues/new.
```