[tests] Fix race condition in the MT5211 mtouch test.
We build for both armv7 and arm64, and the failure we're looking for can
happen with either architecture. The order in which the architectures are
built is random, and mtouch will fail once the first error occurs, which means
that we can't assert that the failure comes from the arm64 build, since the
armv7 build might have been executed first (and triggered the error).

So just don't assert that the expected error message contains the
architecture, which makes the assert valid for both armv7 and armv64.