[builds] Don't reference makefiles from the mono repository when using the mono archive.
Instead of using a makefile in the mono repository to get the list facades to
ship, we list the facade assemblies present in the mono archive instead.

We still have a hardcoded list of facade assemblies elsewhere, and tests that
checks that list against what's shipped, so any changes in the list of
assemblies (unwanted or not) won't go unnoticed.