[launcher] Only set MONO_REGISTRY_PATH if folder exists
This makes VS Mac fail when publishing ASP.NET Core projects to Azure
(see https://devdiv.visualstudio.com/DevDiv/_workitems/edit/896438) because
it calls, when doing so, code that tries to access the registry, when
launched from the Xamarin Installer (or any other Xamarin.Mac app), as
the folder doesn't exist. So, instead of patching it in VS Mac itself
(unsetting the env var), fix it here, so that it's only set if the
app support dir really exists.